% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{bic_from_fit}
\alias{bic_from_fit}
\title{Calculates the BIC score from a fit result.}
\usage{
bic_from_fit(fit_res)
}
\arguments{
\item{fit_res}{A list returned from \code{zi_fit()} or \code{.res_from_model_fits()} with \code{value_only == FALSE}.}
}
\value{
A number, the BIC score.
}
\description{
Calculates the BIC score from a fit result.
}
\details{
Returns \code{bic_score(fit_res$nll, fit_res$n, fit_res$effective_df)}.
}
\examples{
Y <- matrix(rnorm(500) * rbinom(500, 1, 0.7), nrow=50)
res <- zi_fit(Y != 0, Y, 1, 2:5, value_only=FALSE)
bic_from_fit(res)
bic_score(res$nll, res$n, res$effective_df)
}
