% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{bic_score}
\alias{bic_score}
\title{Calculates the BIC score.}
\usage{
bic_score(avgnll, n, numpar)
}
\arguments{
\item{avgnll}{Average negative log likelihood over the sample.}

\item{n}{The sample size.}

\item{numpar}{The degree of freedom.}
}
\value{
A number, the BIC score.
}
\description{
Calculates the BIC score.
}
\details{
Returns \code{2*avgnll+log(n)/n*numpar}.
}
\examples{
Y <- matrix(rnorm(500) * rbinom(500, 1, 0.7), nrow=50)
res <- zi_fit(Y != 0, Y, 1, 2:5, value_only=FALSE)
bic_score(res$nll, res$n, res$effective_df)
bic_from_fit(res)
}
