% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{convert_para}
\alias{convert_para}
\title{Convert parameters in the abk parametrization to pms parametrization or vice versa.}
\usage{
convert_para(paras, pms_to_abk)
}
\arguments{
\item{paras}{A vector of 3 numbers. If \code{pms_to_abk == TRUE}, \code{paras} should contain the log odds and the conditional mean and variance for the Gaussian part. If \code{pms_to_abk == FALSE}, it should contain the \code{a}, \code{b} and \code{k} parameters.}

\item{pms_to_abk}{A logical.}
}
\value{
A vector of 3 numbers.
}
\description{
Gradient of \code{log_dhurdle1d_abk()} (with scalar \code{v} and \code{y}) with respect to \code{a}, \code{b}, \code{k}.
}
\details{
The log density of a sample \code{y} from the 1-d Hurdle model with abk parametrization with respect to the sum of the Lebesgue measure and a point mass at 0 is
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))}
The pms parametrization composes of the log odds (p) and the conditional mean (m) and variance (s) of the Gaussian part, i.e. the density is \eqn{1-p} if \eqn{v == 0}, or \eqn{p*dnorm(x, m, sqrt(s))} if \eqn{v == 1}.
The two can be related as \code{s = 1 / k}, \code{m = b / k}, \code{p = a + b^2/2/k - log(k/2/pi)/2}. Also see Page 7 of McDavid (2019).
}
\examples{
set.seed(1)
convert_para(c(0, 0, 1), TRUE) # (-log(2*pi)/2, 0, 1)
convert_para(c(-log(2*pi)/2, 0, 1), FALSE) # (0, 0, 1)
paras <- c(rnorm(1), rnorm(1), abs(rnorm(1)))
convert_para(convert_para(paras, TRUE), FALSE) - paras
convert_para(convert_para(paras, FALSE), TRUE) - paras

n <- 100
v <- stats::rbinom(100, 1, 0.8)
y <- stats::rnorm(100) * v
a <- paras[1]; b <- paras[2]; k <- paras[3]
tmp <- convert_para(paras, FALSE)
log_odds <- tmp[1]; mu <- tmp[2]; sigmasq <- tmp[3]
prob <- exp(log_odds) / (exp(log_odds)+1)
likelihoods <- numeric(100)
likelihoods[v == 0] <- 1 - prob
likelihoods[v == 1] <- prob * dnorm(y[v == 1], mu, sqrt(sigmasq))
mean(log(likelihoods)) - log_dhurdle1d_abk(v, y, a, b, k)
}
