% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_function.R
\name{CPSS_control}
\alias{CPSS_control}
\title{Returns the graph with desired FDR/freq from a path of CPSS estimates from CPSS_path().}
\usage{
CPSS_control(path, FDR = NULL, freq = NULL, B = 50)
}
\arguments{
\item{path}{A list returned by \code{CPSS_path}}

\item{FDR}{A number between 0 and 1, the desired FDR (false discovery rate); exactly one of \code{FDR} and \code{freq} must be \code{NULL}.}

\item{freq}{A number between 0 and 1, the frequency threshold; exactly one of \code{FDR} and \code{freq} must be \code{NULL}.}

\item{B}{An integer, number of pairs of subsamples to be drawn, should be equal to the \code{B} used when calling \code{CPSS_path()}. Defaults to \code{50}.}
}
\value{
A list associated with the selected estimated DAG.
   \item{graph}{The estimated adjacency matrix.}
   \item{actual_fdr}{The actual FDR controlled.}
   \item{freq}{The frequency threshold for the selected graph using \code{2*B} CPSS estimates.}
}
\description{
Returns the graph with desired FDR/freq from a path of CPSS estimates from CPSS_path().
}
\details{
Returns the densest graph along the CPSS path that has FDR <= \code{FDR} or freq >= \code{freq}. If the resulting graph is empty, returns the sparsest non-empty graph.
Fore more details, see Shah and Samworth (2012) and the documentation on \code{CPSS_fit()} and \code{CPSS_path()}.

Note: The user is expected to call \code{CPSS_fit()}, followed by \code{CPSS_path()}, and then followed by \code{CPSS_control()}.
}
\examples{
m <- 3; n <- 200; B <- 10
adj_mat <- ZiDAG::make_dag(m, mode = "chain")
d <- ZiDAG::gen_zero_dat(seed=1, gen_para="pms", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
fitting_func <- function(indices) return (
    ZiDAG::ziGDS(V=d$V[indices,], Y=d$Y[indices,], parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=1L, "tol"=1e-8, "print_best_degree"=FALSE)))
freq_B <- ZiDAG::CPSS_fit(nrow(d$V), fitting_func, B=B, parallel=FALSE,
    num_cores=min(2, parallel::detectCores()))
cpss_path <- CPSS_path(freq_B, B = B)
CPSS_control(cpss_path, freq=0.5, B=B)
CPSS_control(cpss_path, FDR=0.2, B=B)
}
