% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{DAG_eq}
\alias{DAG_eq}
\title{Returns if the DAGs corresponding to two adjacency matrices are Markov equivalent.}
\usage{
DAG_eq(amat1, amat2)
}
\arguments{
\item{amat1}{The first adjacency matrix.}

\item{amat2}{The second adjacency matrix.}
}
\value{
A logical, whether the DAGs corresponding to two adjacency matrices are Markov equivalent.
}
\description{
Returns if the DAGs corresponding to two adjacency matrices are Markov equivalent.
}
\details{
First tests if \code{amat1} and \code{amat2} are two 2-d matrices of the same dimensions, then test whether they have the same CPDAG using \code{pcalg::dag2cpdag()}.
}
\examples{
# Complete graphs of the same dimension are always Markov equivalent
DAG_eq(make_dag(10, "complete", shuffle=TRUE), make_dag(10, "complete", shuffle=TRUE))
# A v structure not Markov equivalent to a chain graph
DAG_eq(matrix(c(0,1,0, 0,0,0, 0,1,0), nrow=3, byrow=TRUE), make_dag(3, "chain"))
# Reversing a chain graphs gives a Markov equivalent graph
DAG_eq(make_dag(10, "chain"), make_dag(10, "chain")[10:1, 10:1])
}
