% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_mat.R
\name{full_design1}
\alias{full_design1}
\title{Design matrix for pms parametrization given the max degree of the Hurdle polynomial.}
\usage{
full_design1(Vo, Yo, right, V_degree = 1, Y_degree = 1, Y_V_degree = 1)
}
\arguments{
\item{Vo}{A numerical matrix of the same dimension as \code{Yo} indicating if each entry in \code{Yo} is non-zero, i.e. \code{Vo = (Yo != 0)}.}

\item{Yo}{A numerical matrix, the sample matrix for the parent nodes (regressors).}

\item{right}{A string vector of length equal to \code{ncol(Vo)}, the names of the parent nodes.}

\item{V_degree}{A non-negative integer, the degree for the \code{Vo} in the Hurdle polynomial. Defaults to 1.}

\item{Y_degree}{A non-negative integer, the degree for the \code{Yo} in the Hurdle polynomial. Defaults to 1.}

\item{Y_V_degree}{A non-negative integer, the degree for interaction between \code{Vo} and \code{Yo} in the Hurdle polynomial. Defaults to 1. If equal to 1, no interaction will be included (since it would be either a pure \code{V} term or a pure \code{Y} term).}
}
\value{
The design matrix with number of rows equal to \code{nrow(Vo)}.
}
\description{
Design matrix for pms parametrization given the max degree of the Hurdle polynomial.
}
\details{
A Hurdle polynomial in parents \code{Yo} is a polynomial in \code{Yo} and their 0/1 indicators \code{Vo}.
The \code{V_degree} of a term that is a product of some columns of \code{Vo} only is the number of parents that appears in it. For example, \code{V1 * V2 * V3} has \code{V_degree} equal to 3. Note that \code{V1^p} is equal to \code{V1} for any \code{p >= 1} so it does not make sense to include a power.
The \code{Y_degree} of a term that is a product of powers of some columns of \code{Yo} only is the degree of a polynomial in its usual sense. For example, \code{Y1^2 * Y2 * Y3^3} has \code{Y_degree} equal to 2+1+3=6.
The \code{Y_V_degree} of a term that involves both some columns of \code{Vo} and some of \code{Yo} is the sum of the \code{V_degree} of the \code{V} part and the \code{Y_degree} of the \code{Y} part. For example, \code{Y1^2 * V2 * Y3^3 * V4 * V5} has \code{Y_V_degree} equal to 2+1+3+1+1=8.
The function returns a Hurdle polynomial including all possible terms with \code{V_degree}, \code{Y_degree}, \code{Y_V_degree} less than or equal to those specified.
For example, if \code{Vo} and \code{Yo} has two columns and \code{V_degree == 2}, \code{Y_degree == 2}, \code{Y_V_degree == 2}, the function returns a matrix with columns \code{V1}, \code{V2}, \code{V1*V2}, \code{Y1}, \code{Y2}, \code{Y1*Y2}, \code{Y1^2}, \code{Y2^2}, \code{Y1*V2}, \code{Y2*V1}.
Note that terms like \code{V1*Y1} are not included as it is equivalent to \code{Y1}.
Intercepts (a column with all 1s) are also not included.
}
\examples{
Yo <- matrix(rnorm(100) * rbinom(100, 1, 0.7), ncol=2)
full_design1(Yo != 0, Yo, c("1", "2"), V_degree=2, Y_degree=2, Y_V_degree=2)
full_design1(Yo != 0, Yo, c("1", "2"), V_degree=0, Y_degree=0, Y_V_degree=0)
}
