% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_gen.R
\name{gen_zero_dat}
\alias{gen_zero_dat}
\title{Generates a sample zero-inflated data matrix.}
\usage{
gen_zero_dat(
  seed,
  gen_para,
  adj_mat,
  n,
  k_mode = 1,
  min_num = 10,
  therun = 1,
  gen_uniform_degree = 1
)
}
\arguments{
\item{seed}{A number, the seed passed to \code{set.seed()}.}

\item{gen_para}{A string, the parametrization, \code{"abk"} or \code{"pms"}.}

\item{adj_mat}{A square matrix of 0/1s, the adjacency matrix.}

\item{n}{A positive integer, the number of samples.}

\item{k_mode}{One of \code{0}, \code{1}, \code{2}. If \code{k_mode == 1}, the \code{k} (if \code{gen_para == "abk"}) or \code{sigmasq} (if \code{gen_para == "pms"}) parameter in each conditional Hurdle distribution is set to 1. If \code{k_mode == 0}, the \code{k} is set to \code{1/j} (where \code{j} is the topological order of the current node) or \code{sigmasq} is set to \code{j}; if \code{k_mode == 2} they are set in the opposite way.}

\item{min_num}{A positive integer, the minimum number of 0s/1s in each column. If this is not satisfied, this random generator function will be rerun for up to 1000 times.}

\item{therun}{A positive integer for internal use. Defaults to 1.}

\item{gen_uniform_degree}{A positive integer, the maximum degree of the Hurdle polynomials in the conditional Hurdle distributions, defaults to \code{1}. Ignored if \code{parametrization == "abk"}.}
}
\value{
A generated data, containing
    \item{V}{A matrix of 0/1s of dimension \code{ncol(adj_mat) x n}, equal to Y != 0.}
    \item{Y}{A data matrix of dimension \code{ncol(adj_mat) x n}.}
}
\description{
Generates a sample zero-inflated data matrix.
}
\details{
See the simulations section in the paper associated with this package for details.
}
\examples{
x <- gen_zero_dat(1, "abk", make_dag(3, "chain"), 1000)
x <- gen_zero_dat(1, "pms", make_dag(3, "chain"), 1000)
}
