% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_gen.R
\name{generate_one_each_logitp_mu_sigmasq}
\alias{generate_one_each_logitp_mu_sigmasq}
\title{Generates one sample for each triplet of (log_odd, mu) with a fixed sigmasq under the pms parametrization.}
\usage{
generate_one_each_logitp_mu_sigmasq(log_odds, mus, sigmasq)
}
\arguments{
\item{log_odds}{A vector of log odds.}

\item{mus}{A vector of mus, of equal length as \code{log_odds}.}

\item{sigmasq}{A positive number, the sigma squared.}
}
\value{
A vector of the same size as \code{log_odds}.
}
\description{
Generates one sample for each pair of (log_odd, mu) with a fixed sigmasq under the pms parametrization.
}
\details{
Draws one example from the 1-d Hurdle with parameters \code{log_odds[i]}, \code{mus[i]}, \code{sigmasq}, i.e. a mixture of bernoulli and Gaussian with log odds \code{log_odds[i]} for the bernoulli and conditional mean \code{mus[i]} and variance \code{sigmasq} for the Gaussian part, for each \code{i}.
}
\examples{
generate_one_each_logitp_mu_sigmasq(-3:3, 0:6/10, 1)
}
