% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_core.R
\name{grad_a_b_k}
\alias{grad_a_b_k}
\title{Gradient of log_dhurdle1d_abk() (with scalar v and y) with respect to a, b, k.}
\usage{
grad_a_b_k(V, Y, A, B, k)
}
\arguments{
\item{V}{A logical vector, indicating if each entry in \code{Y} is non-zero, i.e. \code{V = (Y != 0)}.}

\item{Y}{A numerical vector of i.i.d. 1-d Hurdle random variables.}

\item{A}{A number or a vector of the same length \code{Y}, the \code{a} parameter(s).}

\item{B}{A number or a vector of the same length \code{Y}, the \code{b} parameter(s).}

\item{k}{A number, the \code{k} parameter.}
}
\value{
A matrix with 3 columns and number of rows equal to the length of \code{Y}.
}
\description{
Gradient of \code{log_dhurdle1d_abk()} (with scalar \code{v} and \code{y}) with respect to \code{a}, \code{b}, \code{k}.
}
\details{
The \code{i}-th row of the returned matrix is the gradient of \code{log_dhurdle1d_abk(V[i], Y[i], A[i], B[i], k)} with respect to \code{A[i]}, \code{B[i]}, \code{k} if \code{A} and \code{B} are vectors, or with respect to \code{A} and \code{B} if they are scalars.
}
\examples{
if (!requireNamespace("numDeriv", quietly = TRUE))
   stop("Please install package \"numDeriv\".")
set.seed(1)
n <- 100
V <- rbinom(n, 1, 0.8)
Y <- rnorm(n) * V
A <- rnorm(n)
B <- rnorm(n)
k <- abs(rnorm(1))
grad_mat <- grad_a_b_k(V, Y, A, B, k)
numer_grad <- t(sapply(1:n,
   function(i){numDeriv::grad(function(x){
      log_dhurdle1d_abk(V[i], Y[i], x[1], x[2], x[3])}, c(A[i], B[i], k))}))
max(abs(grad_mat - numer_grad))

a <- rnorm(1)
b <- rnorm(1)
grad_mat <- grad_a_b_k(V, Y, a, b, k)
numer_grad <- t(sapply(1:n,
   function(i){numDeriv::grad(function(x){
      log_dhurdle1d_abk(V[i], Y[i], x[1], x[2], x[3])}, c(a, b, k))}))
max(abs(grad_mat - numer_grad))
}
