% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{log_dhurdle1d_abk}
\alias{log_dhurdle1d_abk}
\title{Calculates the mean log probability of a vector of i.i.d. 1-d Hurdle random variables under the abk parametrization.}
\usage{
log_dhurdle1d_abk(V, Y, a, b, k)
}
\arguments{
\item{V}{A logical vector, indicating if each entry in \code{Y} is non-zero, i.e. \code{V = (Y != 0)}.}

\item{Y}{A numerical vector of i.i.d. 1-d Hurdle random variables.}

\item{a}{A number, the \code{a} parameter.}

\item{b}{A number, the \code{b} parameter.}

\item{k}{A number, the \code{k} parameter.}
}
\value{
A number, the mean log probability.
}
\description{
Calculates the mean log probability of a vector of i.i.d. 1-d Hurdle random variables under the abk parametrization.
}
\details{
The log probability of a sample \code{y} from the 1-d Hurdle model with abk parametrization with respect to the sum of the Lebesgue measure and a point mass at 0 is
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))}
where \code{v = (y != 0)}. The function calculates the mean of log probabilities over a vector of i.i.d samples.
}
\examples{
y <- rhurdle1d_abk(n=100, a=1, b=2, k=3)
log_dhurdle1d_abk(y != 0, y, a=1, b=2, k=3)
}
