% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{log_dhurdle1d_pms}
\alias{log_dhurdle1d_pms}
\title{Calculates the mean log probability of a vector of i.i.d. 1-d Hurdle random variables under the pms parametrization.}
\usage{
log_dhurdle1d_pms(V, Y, p, mu, sigmasq)
}
\arguments{
\item{V}{A logical vector, indicating if each entry in \code{Y} is non-zero, i.e. \code{V = (Y != 0)}.}

\item{Y}{A numerical vector of i.i.d. 1-d Hurdle random variables.}

\item{p}{A number, the \code{p} parameter.}

\item{mu}{A number, the \code{mu} parameter.}

\item{sigmasq}{A number, the \code{sigmasq} parameter.}
}
\value{
A number, the mean log probability.
}
\description{
Calculates the mean log probability of a vector of i.i.d. 1-d Hurdle random variables under the pms parametrization.
}
\details{
The log probability of a sample \code{y} from the 1-d Hurdle model with pms parametrization with respect to the sum of the Lebesgue measure and a point mass at 0 is
\eqn{\log(1-p)}{log(1-p)} if it is equal to 0, or \eqn{\log(p)-(y-mu)^2/2/sigmasq}{log(p)-(y-mu)^2/2/sigmasq} otherwise.
The function calculates the mean of log probabilities over a vector of i.i.d samples.
}
\examples{
y <- rhurdle1d_pms(n=100, log_odds=1.2, mu=2.3, sigmasq=3.4)
log_dhurdle1d_pms(y != 0, y, p=exp(1.2)/(exp(1.2)+1), mu=2.3, sigmasq=3.4)
}
