% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{log_dhurdle_vec_abk}
\alias{log_dhurdle_vec_abk}
\title{Calculates the mean log probability of a vector of 1-d Hurdle random variables under the abk parametrization, each variable with different a and b and shared k.}
\usage{
log_dhurdle_vec_abk(V, Y, A, B, k)
}
\arguments{
\item{V}{A logical vector, indicating if each entry in \code{Y} is non-zero, i.e. \code{V = (Y != 0)}.}

\item{Y}{A numerical vector of i.i.d. 1-d Hurdle random variables.}

\item{A}{A numerical vector of the same size as \code{V} and \code{Y}, the \code{a} parameters for each sample from \code{Y}.}

\item{B}{A numerical vector of the same size as \code{V} and \code{Y}, the \code{b} parameters for each sample from \code{Y}.}

\item{k}{A positive number, the common \code{k} parameter for all samples from \code{Y}.}
}
\value{
A number, the mean log probability.
}
\description{
Calculates the mean log probability of a vector of 1-d Hurdle random variables under the abk parametrization, each variable with different \code{a} and \code{b} and shared \code{k}.
}
\details{
The log probability of a sample \code{y} from the 1-d Hurdle model with abk parametrization with respect to the sum of the Lebesgue measure and a point mass at 0 is
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))}
where \code{v = (y != 0)}. The function calculates the mean of the log probabilities over a vector of samples with different \code{a} and \code{b} parameters and the same \code{k}.
This is used for the multivariate case where in the conditional 1-d Hurdle model the \code{a} and \code{b} are functions in other (parent) nodes/variables.
}
\examples{
set.seed(1)
A <- rnorm(100)
B <- rnorm(100)
y <- generate_one_each_a_b_k(A=A, B=B, k=2)
log_dhurdle_vec_abk(y != 0, y, A=A, B=B, k=2)
}
