% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{logsumexp_vec}
\alias{logsumexp_vec}
\title{Calculates mean(log(1+exp(xs))) of a vector xs.}
\usage{
logsumexp_vec(xs)
}
\arguments{
\item{xs}{A vector of random numbers.}
}
\value{
A number.
}
\description{
Calculates \code{mean(log(1+exp(xs)))} of a vector \code{xs}.
}
\details{
Numerically stable calculation of \code{mean(log(1+exp(xs)))}.
}
\examples{
logsumexp_vec(10^(-5:5))
logsumexp_vec(rep(0, 5))
}
