% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_gen.R
\name{make_dag}
\alias{make_dag}
\title{Returns the adjacency matrix of a chain, complete, or a lattice DAG.}
\usage{
make_dag(m, mode, nrows = 0, shuffle = FALSE, seed = NULL)
}
\arguments{
\item{m}{A positive integer, the dimension.}

\item{mode}{One of \code{"chain"}, \code{"complete"}, \code{"lattice"}.}

\item{nrows}{A positive integer that divides \code{m}, used for \code{mode == "lattice"} as the number of rows in the graph.}

\item{shuffle}{A logical, indicates whether the nodes \code{1:m} should be shuffled. Defaults to \code{FALSE}.}

\item{seed}{A number, the seed passed to \code{set.seed()} if \code{shuffle == TRUE}.}
}
\value{
The \code{m x m} adjacency matrix of the DAG.
}
\description{
Returns the adjacency matrix of a chain, complete, or a lattice DAG.
}
\details{
A chain graph has the form 1 -> 2 -> 3 -> ... -> m.
A complete graph has an edge from \code{i} to \code{j} for any \code{1 <= i < j <= m}.
A lattice graph first places the nodes in a \code{nrows x (m/nrows)} matrix, row by row. That is, the \code{i}-th node is on the \code{i \%/\% (m/nrows) + 1}-th row and the \code{i \%\% (m/nrows)}-th column (or the \code{(m/nrows)}-th column if \code{i} divides it). Then add an edge between any two adjacent nodes, pointing from the left to the right or from the node above to the node below.
If \code{shuffle == TRUE}, the rows/columns of the returning adjacency matrix are simultaneously shuffled.
}
\examples{
make_dag(10, "chain", shuffle=FALSE)
make_dag(10, "chain", shuffle=TRUE)
make_dag(10, "complete", shuffle=FALSE)
make_dag(10, "complete", shuffle=TRUE)
make_dag(12, "lattice", nrows=3, shuffle=FALSE)
make_dag(12, "lattice", nrows=3, shuffle=TRUE)
}
