% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_core.R
\name{mle1d_abk}
\alias{mle1d_abk}
\title{Fits MLE for 1-d Hurdle model under the abk parametrization.}
\usage{
mle1d_abk(V, Y, left, value_only = TRUE)
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{left}{An integer between 1 and \code{ncol(Y)}. The index of the variable to be fit.}

\item{value_only}{If \code{TRUE}, returns the minimized negative log likelihood only. Defaults to \code{TRUE}.}
}
\value{
If \code{value_only == TRUE}, returns the minimized negative log likelihood only.
Otherwise, returns
  \item{nll}{A number, the minimized negative log likelihood.}
  \item{par}{A vector of 3 numbers, the fitted a, b, and k parameters.}
  \item{n}{An integer, the sample size.}
  \item{effective_df}{Number 3, the effective degree of freedom.}
}
\description{
Fits MLE for 1-d Hurdle model under the abk parametrization.
}
\details{
Fits an MLE for the 1-d Hurdle model to \code{Y[,left]} under the abk parametrization.
}
\examples{
Y <- matrix(rhurdle1d_abk(n=1e4, a=1.2, b=2.3, k=1.9), ncol=1)
V <- Y != 0
mle1d_abk(V, Y, 1, TRUE)
mle1d_abk(V, Y, 1, FALSE)
}
