% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{rand_init_dags}
\alias{rand_init_dags}
\title{Randomly generates DAGs of specified sparsity.}
\usage{
rand_init_dags(m, init_spars, init_eachs)
}
\arguments{
\item{m}{An integer, the number of nodes.}

\item{init_spars}{A vector of numbers between 0 and 1, inclusive. It is assumed that \code{as.integer(m*(m-1)/2*(1-init_spars))} does not contain duplicates.}

\item{init_eachs}{A positive integer or a vector of positive integers of the same length as \code{init_spars}.}
}
\value{
A list of adjacency matrices each of shape \code{m} by \code{m}.
}
\description{
Randomly generates DAGs of specified sparsity.
}
\details{
Randomly generates DAGs of specified sparsity.
For each \code{spar} in \code{init_spars}, the function generates a complete graph, then randomly remove \code{as.integer(m*(m-1)/2*(1-spar))} edges. Finally, the ordering of the \code{m} nodes are randomized.
For each \code{spar = init_spars[i]} this is repeated \code{init_eachs} (if integer) or \code{init_eachs[i]} (if \code{init_eachs} is a vector of integers) times; duplicate graphs are avoided and at most 200 attempts will be made when generating each new graph.
Note that for \code{spar = 0} there will only be one empty graph returned.
}
\examples{
rand_init_dags(4, c(0, 0.25, 0.5, 0.75, 1), 5)
rand_init_dags(10, c(0, 0.25, 0.5, 0.75, 1), 5)
}
