% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_gen.R
\name{rhurdle1d_abk}
\alias{rhurdle1d_abk}
\title{Generates samples from the abk parametrization.}
\usage{
rhurdle1d_abk(n, a, b, k)
}
\arguments{
\item{n}{The number of observations to be drawn.}

\item{a}{The \code{a} parameter.}

\item{b}{The \code{b} parameter.}

\item{k}{The \code{k} parameter.}
}
\value{
A vector of size \code{n}.
}
\description{
Generates samples from the abk parametrization.
}
\details{
Generates samples from the Hurdle 1d distribution, whose log density with abk parametrization with respect to the sum of the Lebesgue measure and a point mass at 0 is
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))}
where \code{v = (y != 0)}.
}
\examples{
set.seed(1)
a <- stats::rnorm(1)
b <- stats::rnorm(1)
k <- abs(stats::rnorm(1))
x <- rhurdle1d_abk(1e5, a, b, k)
c(mean(x != 0), 1 - 1 / (1 + sqrt(2*pi/k) * exp(a+b^2/(2*k))))
hist(x[x != 0], freq=FALSE, breaks=200)
curve(exp(b*x-x^2*k/2)/(1+sqrt(2*pi/k)*exp(a+b^2/(2*k))),
   from=min(x), to=max(x), add=TRUE, col="red")
}
