% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_gen.R
\name{rhurdle1d_pms}
\alias{rhurdle1d_pms}
\title{Generates samples from the pms parametrization.}
\usage{
rhurdle1d_pms(n, log_odds, mu, sigmasq)
}
\arguments{
\item{n}{The number of observations to be drawn.}

\item{log_odds}{The log odds, the \code{p} parameter.}

\item{mu}{The conditional mean of the Gaussian part, the \code{m} parameter.}

\item{sigmasq}{The conditional variance of the Gaussian part, the \code{s} parameter.}
}
\value{
A vector of size \code{n}.
}
\description{
Generates samples from the pms parametrization.
}
\details{
Generates samples from the Hurdle 1d distribution, with probability \eqn{1-p} for \eqn{y == 0}, and density \eqn{p*dnorm(x, m, sqrt(s))} for \eqn{y != 0}.
}
\examples{
set.seed(1)
log_odds <- stats::rnorm(1)
mu <- stats::rnorm(1)
sigmasq <- abs(stats::rnorm(1))
x <- rhurdle1d_pms(1e5, log_odds, mu, sigmasq)
c(mean(x != 0), exp(log_odds) / (exp(log_odds) + 1))
hist(x[x != 0], freq=FALSE, breaks=200)
curve(dnorm(x, mu, sqrt(sigmasq)), from=min(x), to=max(x), add=TRUE, col="red")
}
