% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{sum_a}
\alias{sum_a}
\title{Calculate the a parameter in the conditional Hurdle density of the child node given the parents for one sample, assuming a is linear in the parents and their indicators.}
\usage{
sum_a(aa, vo, yo)
}
\arguments{
\item{aa}{A numerical vector of length \code{2*length(yo)+1}. See details.}

\item{vo}{A numerical vector of the same length as \code{yo} indicating if each entry in \code{yo} is non-zero, i.e. \code{vo = (yo != 0)}.}

\item{yo}{A numerical vector, a sample for the parent nodes (regressors).}
}
\value{
A number, the mean log probability.
}
\description{
Calculate the \code{a} parameter in the conditional Hurdle density of the child node given the parents for one sample, assuming \code{a} is linear in the parents and their indicators.
}
\details{
Assuming \code{aa = c(a1, a2, a3)} where \code{a1} is a scalar, \code{a2} and \code{a3}, the function returns \code{a1 + sum(a2*vo) + sum(a3*yo)}.
This is used for calculating the \code{a} parameter of the conditional Hurdle density of the child node given the parent nodes, assuming the \code{a} parameter is linear in the parents and their indicators.
}
\examples{
set.seed(1)
vo <- stats::rbinom(100, 1, 0.8)
yo <- stats::rnorm(100) * vo
sum_a(stats::rnorm(201), vo, yo)
}
