% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{sum_A_mat}
\alias{sum_A_mat}
\title{Calculate the a parameter in the conditional Hurdle density of the child node given the parents for multiple samples, assuming a is linear in the parents and their indicators.}
\usage{
sum_A_mat(aa, Vo, Yo)
}
\arguments{
\item{aa}{A numerical vector of length \code{2*ncol(Yo)+1}. See details.}

\item{Vo}{A numerical vector of the same dimension as \code{Yo} indicating if each entry in \code{Yo} is non-zero, i.e. \code{Vo = (Yo != 0)}.}

\item{Yo}{A numerical vector, a sample for the parent nodes (regressors).}
}
\value{
A number, the mean log probability.
}
\description{
Calculate the \code{a} parameter in the conditional Hurdle density of the child node given the parents for multiple samples, assuming \code{a} is linear in the parents and their indicators.
}
\details{
Matrix version of \code{sum_a()}. See examples.
}
\examples{
set.seed(1)
n <- 100; p_others <- 10
Vo <- matrix(stats::rbinom(n*p_others, 1, 0.8), nrow=n, ncol=p_others)
Yo <- matrix(stats::rnorm(n*p_others) * Vo, nrow=n, ncol=p_others)
aa <- rnorm(2*p_others+1)
sum_A_mat(aa, Vo, Yo) - sapply(1:n, function(i){sum_a(aa, Vo[i,], Yo[i,])})
}
