% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{sum_b}
\alias{sum_b}
\title{Calculate the b parameter in the conditional Hurdle density of the child node given the parents for one sample, assuming b is linear in the parents and their indicators.}
\usage{
sum_b(bb, vo, yo, minus_Y = TRUE)
}
\arguments{
\item{bb}{A numerical vector of length \code{2*length(yo)+1}. See details.}

\item{vo}{A numerical vector of the same length as \code{yo} indicating if each entry in \code{yo} is non-zero, i.e. \code{vo = (yo != 0)}.}

\item{yo}{A numerical vector, a sample for the parent nodes (regressors).}

\item{minus_Y}{A logical, see details.}
}
\value{
A number, the mean log probability.
}
\description{
Calculate the \code{b} parameter in the conditional Hurdle density of the child node given the parents for one sample, assuming \code{b} is linear in the parents and their indicators.
}
\details{
Assuming \code{bb = c(b1, b2, b3)} where \code{b1} is a scalar, \code{b2} and \code{b3}, the function returns \code{b1 + sum(b2*vo) - sum(b3*yo)} if \code{minus_Y == TRUE}, otherwise \code{b1 + sum(b2*vo) + sum(b3*yo)}.
This is used for calculating the \code{a} parameter of the conditional Hurdle density of the child node given the parent nodes, assuming the \code{b} parameter is linear in the parents and their indicators.
}
\examples{
set.seed(1)
vo <- stats::rbinom(100, 1, 0.8)
yo <- stats::rnorm(100) * vo
sum_b(stats::rnorm(201), vo, yo, minus_Y=TRUE)
}
