% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dist.R
\name{sum_B_mat}
\alias{sum_B_mat}
\title{Calculate the b parameter in the conditional Hurdle density of the child node given the parents for multiple samples, assuming b is linear in the parents and their indicators.}
\usage{
sum_B_mat(bb, Vo, Yo, minus_Y = TRUE)
}
\arguments{
\item{bb}{A numerical vector of length \code{2*ncol(Yo)+1}. See details.}

\item{Vo}{A numerical vector of the same dimension as \code{Yo} indicating if each entry in \code{Yo} is non-zero, i.e. \code{Vo = (Yo != 0)}.}

\item{Yo}{A numerical vector, a sample for the parent nodes (regressors).}

\item{minus_Y}{A logical, see details in \code{sum_b()}.}
}
\value{
A number, the mean log probability.
}
\description{
Calculate the \code{b} parameter in the conditional Hurdle density of the child node given the parents for multiple samples, assuming \code{b} is linear in the parents and their indicators.
}
\details{
Matrix version of \code{sum_b()}. See examples.
}
\examples{
set.seed(1)
n <- 100; p_others <- 10
Vo <- matrix(stats::rbinom(n*p_others, 1, 0.8), nrow=n, ncol=p_others)
Yo <- matrix(stats::rnorm(n*p_others) * Vo, nrow=n, ncol=p_others)
bb <- rnorm(2*p_others+1)
sum_B_mat(bb, Vo, Yo) - sapply(1:n, function(i){sum_b(bb, Vo[i,], Yo[i,])})
}
