% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_core.R
\name{zi_fit}
\alias{zi_fit}
\title{Fits a Hurdle polynomial with either abk or pms parametrization.}
\usage{
zi_fit(
  V,
  Y,
  left,
  right,
  parametrization = "pms",
  value_only = TRUE,
  control = list()
)
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{left}{An integer between 1 and \code{ncol(Y)}. The index of the variable to be fit.}

\item{right}{A vector of integers between 1 and \code{ncol(Y)} different from \code{left}. Indices of the "regressors".}

\item{parametrization}{A string, either \code{"abk"} (canonical) or \code{"pms"} (moment).}

\item{value_only}{If \code{TRUE}, returns the minimized negative log likelihood only. Defaults to \code{TRUE}.}

\item{control}{A list of arguments to be passed to \code{zi_fit_abk()}, \code{zi_fit_pms()} and \code{zi_fit_pms_choose_degree()}.}
}
\value{
If \code{value_only == TRUE}, returns the minimized negative log likelihood only. Otherwise, returns
  \item{nll}{A number, the minimized negative log likelihood.}
  \item{par}{A vector of length \code{4*length(right)+3}, the fitted parameters, in the other of: the intercept for the \code{a} (a scalar), linear coefficients on \code{V[,right]} for \code{a}, linear coefficients on \code{Y[,right]} for \code{a}, the intercept for the \code{b} (a scalar), linear coefficients on \code{V[,right]} for \code{b}, linear coefficients on \code{Y[,right]} for \code{b}.}
  \item{n}{An integer, the sample size.}
  \item{effective_df}{\code{4*length(right)+3}, the effective degree of freedom.}
}
\description{
Fits a Hurdle polynomial with either abk or pms parametrization.
}
\details{
For \code{"abk"} parametrization, consult \code{?zi_fit_abk} for \code{control}.
For \code{"pms"} parametrization, consult \code{?zi_fit_pms} for model fitting using Hurdle polynomials of an exact degree, or \code{?zi_fit_pms_choose_degree} for automatically determining the degree given a maximum degree allowed.

If \code{parametrization == "abk"}, calls \code{zi_fit_abk(V, Y, left, right, value_only)} with other arguments set to those in the \code{control} list.
If \code{parametrization == "pms" && is.null(control[["max_uniform_degree"]])}, calls \code{zi_fit_pms(V, Y, left, right, value_only)} with other arguments set to those in the \code{control} list.
If \code{parametrization == "pms" && !is.null(control[["max_uniform_degree"]])}, calls \code{zi_fit_pms_choose_degree(V, Y, left, right, value_only)} with other arguments set to those in the \code{control} list.
For the list of arguments for \code{zi_fit_abk()}, \code{zi_fit_pms()} and \code{zi_fit_pms_choose_degree()}, please refer to their documentation.
}
\examples{
m <- 3; n <- 1000
adj_mat <- make_dag(m, "complete")
dat <- gen_zero_dat(1, "abk", adj_mat, n, k_mode=1, min_num=10, gen_uniform_degree=1)
zi_fit(dat$V, dat$Y, 3, 1:2, "abk", TRUE, list(use_C=FALSE, maxit=1000, runs=2, report=0))
zi_fit(dat$V, dat$Y, 3, 1:2, "abk", TRUE, list(use_C=TRUE, maxit=1000, runs=2, report=0))
zi_fit(dat$V, dat$Y, 3, 1:2, "abk", FALSE, list(use_C=FALSE, maxit=1000, runs=2, report=0))
zi_fit(dat$V, dat$Y, 3, 1:2, "abk", FALSE, list(use_C=TRUE, maxit=1000, runs=2, report=0))

dat <- gen_zero_dat(1, "pms", adj_mat, 1000, k_mode=1, min_num=10, gen_uniform_degree=1)
zi_fit(dat$V, dat$Y, 3, 1:2, "pms", value_only=TRUE,
    list(p_V_degree=2, p_Y_degree=2, p_Y_V_degree=2, mu_V_degree=2, mu_Y_degree=2))
zi_fit(dat$V, dat$Y, 3, 1:2, "pms", value_only=FALSE,
    list(p_V_degree=2, p_Y_degree=2, p_Y_V_degree=2, mu_V_degree=2, mu_Y_degree=2))

zi_fit(dat$V, dat$Y, 3, 1:2, "pms", value_only=TRUE,
    list(max_uniform_degree=2L, print_best_degree = TRUE))
zi_fit(dat$V, dat$Y, 3, 1:2, "pms", value_only=FALSE,
    list(max_uniform_degree=2L, print_best_degree = TRUE))
}
