% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_core.R
\name{zi_fit_C}
\alias{zi_fit_C}
\title{Fits the Hurdle model assuming linear abk parametrization using C code.}
\usage{
zi_fit_C(
  V,
  Y,
  left,
  right,
  maxit = 200,
  runs = 1,
  value_only = TRUE,
  verbosity = 0,
  step_size = 0.1,
  lm_tol = 0.01,
  epsabs = 1e-10,
  maxsize = 1e-10,
  method = 5
)
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{left}{An integer between 1 and \code{ncol(Y)}. The index of the variable to be fit.}

\item{right}{A vector of integers between 1 and \code{ncol(Y)} different from \code{left}. Indices of the "regressors".}

\item{maxit}{An integer, the maximum number of integers, argument to \code{stats::optim}.}

\item{runs}{A positive integer, number of reruns; if larger than \code{1}, \code{stats::optim} will be called multiple times with different initial values, and the fit that maximizes the log likelihood will be used.}

\item{value_only}{If \code{TRUE}, returns the minimized negative log likelihood only. Defaults to \code{TRUE}.}

\item{verbosity}{A positive integer, verbosity level. If set to 0 no output is print during estimation.}

\item{step_size}{A double, the step size of the first trial step. Defaults to 0.1.}

\item{lm_tol}{A double, accuracy of the line minimization.}

\item{epsabs}{A double, accuracy of the minimization.}

\item{maxsize}{A double, the final size of the simplex.}

\item{method}{A positive integer indicating the method to use. 5 is recommended and default.
\describe{
  \item{0}{Fletcher-Reeves conjugate gradient}
  \item{1}{Polak-Ribiere conjugate gradient}
  \item{2}{Vector Broyden-Fletcher-Goldfarb-Shanno method}
  \item{3}{Steepest descent algorithm}
  \item{4}{Nelder-Mead simplex}
  \item{5}{Vector Broyden-Fletcher-Goldfarb-Shanno method version 2}
  \item{6}{Simplex algorithm of Nelder and Mead version 2}
  \item{7}{Simplex algorithm of Nelder and Mead: random initialization}
}}
}
\value{
If \code{value_only == TRUE}, returns the minimized negative log likelihood only. Otherwise, returns
  \item{nll}{A number, the minimized negative log likelihood.}
  \item{par}{A vector of length \code{4*length(right)+3}, the fitted parameters, in the other of: the intercept for the \code{a} (a scalar), linear coefficients on \code{V[,right]} for \code{a}, linear coefficients on \code{Y[,right]} for \code{a}, the intercept for the \code{b} (a scalar), linear coefficients on \code{V[,right]} for \code{b}, linear coefficients on \code{Y[,right]} for \code{b}.}
  \item{n}{An integer, the sample size.}
  \item{effective_df}{\code{4*length(right)+3}, the effective degree of freedom.}
}
\description{
Fits the Hurdle model assuming linear abk parametrization using C code.
}
\details{
Fits the Hurdle model assuming linear abk parametrization, where \code{Y[,left]} conditional on \code{Y[,right]} is a 1-d Hurdle model with respect to the sum of the Lebesgue measure and a point mass at 0 with density
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))},
with \code{a} and \code{b} both linear functions in \code{V[,right]} and \code{Y[,right]}.
}
\examples{
m <- 3; n <- 1000
adj_mat <- make_dag(m, "complete")
dat <- gen_zero_dat(1, "abk", adj_mat, n, k_mode=1, min_num=10, gen_uniform_degree=1)
zi_fit_C(dat$V, dat$Y, 3, 1:2, maxit=1000, runs=2, value_only=TRUE, verbosity=0)
zi_fit_C(dat$V, dat$Y, 3, 1:2, maxit=1000, runs=2, value_only=FALSE, verbosity=0)
}
