% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_fit_pms.R
\name{zi_fit_lm}
\alias{zi_fit_lm}
\title{Performs unpenalized/ridge logistic/linear regression.}
\usage{
zi_fit_lm(
  y,
  X,
  binomial,
  intercept = TRUE,
  tol = 1e-08,
  maxit = 1e+05,
  seed = NULL,
  penalize_decider = function(X) {     ncol(X) >= nrow(X)/2 },
  pen_factors = NULL,
  nfits = 10,
  runs = 2,
  CV_or_BIC = FALSE
)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The design matrix.}

\item{binomial}{A logical. Fits a logistic regression if \code{TRUE}, otherwise a linear regression.}

\item{intercept}{A logical, whether to include an intercept in the regression. Defaults to \code{TRUE}.}

\item{tol}{A number, tolerance. Defaults to \code{1e-8}. Passed to \code{stats::glm()} for penalized logistic regressions, or as the \code{thresh} argument to \code{glmnet::glmnet()} for both logistic and linear regressions if penalized.}

\item{maxit}{An integer, the maximum number of iterations. Defaults to \code{100000}. Passed to \code{stats::glm()} for penalized logistic regressions, or to \code{glmnet::glmnet()} for both logistic and linear regressions if penalized.}

\item{seed}{The seed, passed to \code{set.seed()} before model fitting if not \code{NULL}. Defaults to \code{NULL}.}

\item{penalize_decider}{A logical or a function that takes \code{X} and returns a logical. Defaults to \code{function(X){ncol(X)>=nrow(X)/2}}. Used to decide whether to use penalized l2 (ridge) regression (if \code{TRUE}). Note that if the fits for the unpenalized regression are almost perfect, a penalized regression will be automatically used.}

\item{pen_factors}{A vector of non-negative numbers, defaults to \code{NULL}. Passed as the \code{penalty.factor} argument to \code{glmnet::glmnet()}. If \code{intercept == TRUE}, a \code{0} will be prepended.}

\item{nfits}{A positive integer, defaults to \code{10}. Used for penalized regressions, as number of folds if \code{CV_BIC == TRUE} (\code{nfits} argument to \code{glmnet::cv.glmnet()}, with \code{nlambda} set to \code{100}), or the number of lambdas if \code{BIC == FALSE} (as the \code{nlambda} argument to \code{glmnet::glmnet()}).}

\item{runs}{A positive integer, the number of reruns. The fit with the maximum likelihood will be returned. Defaults to \code{2}.}

\item{CV_or_BIC}{A logical, whether to use cross validation or BIC to choose the model from the path of penalized fits. Defaults to \code{FALSE}.}
}
\value{
A list of following elements for the fitted model:
    \item{coefficients}{The coefficients on columns of \code{X}. The coefficient on the intercept is prepended if \code{intercept == TRUE}.}
    \item{logLik}{The log likelihood.}
    \item{sigmasq}{The estimated variance (sigma squared). \code{NULL} if \code{binomial == FALSE}.}
    \item{df}{The effective degree of freedom.}
}
\description{
Performs unpenalized/ridge logistic/linear regression.
}
\details{
See description of each argument.
}
\examples{
m <- 5; n <- 100;
adj_mat <- ZiDAG::make_dag(m, mode = "chain")
d <- ZiDAG::gen_zero_dat(seed=1, gen_para="pms", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
X <- full_design1(d$V[,2:m], d$Y[,2:m], right=paste(2:m),
         V_degree = 2, Y_degree = 2, Y_V_degree = 2)
zi_fit_lm(d$V[,1], X, TRUE)
zi_fit_lm(d$V[,1], X, penalize_decider=TRUE, TRUE, nfits=5, runs=1)
zi_fit_lm(d$Y[,1], X, FALSE)
zi_fit_lm(d$Y[,1], X, FALSE, penalize_decider=function(X){nrow(X)<100*ncol(X)}, nfits=5, runs=1)
}
