% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_core.R
\name{zi_fit_R}
\alias{zi_fit_R}
\title{Fits the Hurdle model assuming linear abk parametrization using \code{stats::optim}.}
\usage{
zi_fit_R(
  V,
  Y,
  left,
  right,
  maxit = 200,
  tol = 1e-08,
  runs = 1,
  value_only = TRUE,
  report = 10
)
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{left}{An integer between 1 and \code{ncol(Y)}. The index of the variable to be fit.}

\item{right}{A vector of integers between 1 and \code{ncol(Y)} different from \code{left}. Indices of the "regressors".}

\item{maxit}{An integer, the maximum number of integers, argument to \code{stats::optim}.}

\item{tol}{A positive number, the tolerance passed as the \code{pgtol} argument to \code{stats::optim}.}

\item{runs}{A positive integer, number of reruns; if larger than \code{1}, \code{stats::optim} will be called multiple times with different initial values, and the fit that maximizes the log likelihood will be used.}

\item{value_only}{If \code{TRUE}, returns the minimized negative log likelihood only. Defaults to \code{TRUE}.}

\item{report}{An integer indicating verbosity, argument to \code{stats::optim}. If set to 1 no output is print during estimation.}
}
\value{
If \code{value_only == TRUE}, returns the minimized negative log likelihood only. Otherwise, returns
  \item{nll}{A number, the minimized negative log likelihood.}
  \item{par}{A vector of length \code{4*length(right)+3}, the fitted parameters, in the other of: the intercept for the \code{a} (a scalar), linear coefficients on \code{V[,right]} for \code{a}, linear coefficients on \code{Y[,right]} for \code{a}, the intercept for the \code{b} (a scalar), linear coefficients on \code{V[,right]} for \code{b}, linear coefficients on \code{Y[,right]} for \code{b}.}
  \item{n}{An integer, the sample size.}
  \item{effective_df}{\code{4*length(right)+3}, the effective degree of freedom.}
}
\description{
Fits the Hurdle model assuming linear abk parametrization using \code{stats::optim}.
}
\details{
Fits the Hurdle model assuming linear abk parametrization, where \code{Y[,left]} conditional on \code{Y[,right]} is a 1-d Hurdle model with respect to the sum of the Lebesgue measure and a point mass at 0 with density
\eqn{av+by-y^2k/2-\log(1+\sqrt(2\pi/k)\exp(a+b^2/(2k)))}{a*v+b*y-y^2*k/2-log(1+sqrt(2pi/k)*exp(a+b^2/(2k)))},
with \code{a} and \code{b} both linear functions in \code{V[,right]} and \code{Y[,right]}.
}
\examples{
m <- 3; n <- 1000
adj_mat <- make_dag(m, "complete")
dat <- gen_zero_dat(1, "abk", adj_mat, n, k_mode=1, min_num=10, gen_uniform_degree=1)
zi_fit_R(dat$V, dat$Y, 3, 1:2, maxit=1000, runs=2, value_only=TRUE, report=0)
zi_fit_R(dat$V, dat$Y, 3, 1:2, maxit=1000, runs=2, value_only=FALSE, report=0)
}
