% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{ziGDS}
\alias{ziGDS}
\title{Greedy DAG search for DAGs for zero-inflated data based on Hurdle conditionals.}
\usage{
ziGDS(
  V,
  Y,
  parametrization,
  fixedGaps = NULL,
  maxDegree = Inf,
  maxInDegree = Inf,
  init = NULL,
  verbose = FALSE,
  control = list()
)
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{parametrization}{A string, either \code{"abk"} (canonical) or \code{"pms"} (moment).}

\item{fixedGaps}{A logical square matrix of number of rows/columns equal to number of columns of \code{Y}. Similar to \code{pcalg::gds()}, if its \code{[i,j]} is \code{TRUE}, the result is guaranteed to have no edge between nodes \code{i} and \code{j}.}

\item{maxDegree}{A positive integer, the maximum degree of a node.}

\item{maxInDegree}{A positive integer, the maximum in-degree of a node.}

\item{init}{An adjacency matrix of number of rows/columns equal to number of columns of \code{Y}. Defaults to \code{NULL} and used as the initial graph. If not specified, the empty graph is used for initialization.}

\item{verbose}{A logical, whether to print intermediate steps.}

\item{control}{A list passed to \code{zi_fit()}. Please consult \code{?zi_fit}.}
}
\value{
An adjacency matrix.
}
\description{
Greedy DAG search for DAGs for zero-inflated data based on Hurdle conditionals.
}
\details{
Performs greedy DAG search for DAGs for zero-inflated data based on Hurdle conditionals.
See Chickering (2003) or \code{?pcalg::gds} for details. However, unlike the Gaussian case in their implementation that returns an equivalence class of DAGs, the DAG estimated by this function is exact.
}
\examples{
m <- 4; n <- 1000
adj_mat <- ZiDAG::make_dag(m, mode = "chain", shuffle=FALSE)
d <- ZiDAG::gen_zero_dat(seed=1, gen_para="pms", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
est <- ZiDAG::ziGDS(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=1L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est

d <- ZiDAG::gen_zero_dat(seed=1, gen_para="abk", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
est <- ZiDAG::ziGDS(V=d$V, Y=d$Y, parametrization="abk", verbose=FALSE,
    control=list(use_C=TRUE, maxit=1000, runs=2, report=0))
adj_mat == est

est <- ZiDAG::ziGDS(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=1L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est

est <- ZiDAG::ziGDS(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=2L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est
}
