% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_est_graph.R
\name{ziSIMY}
\alias{ziSIMY}
\title{Exhaustive search using BIC for DAGs for zero-inflated data based on Hurdle conditionals.}
\usage{
ziSIMY(V, Y, parametrization, verbose = FALSE, control = list())
}
\arguments{
\item{V}{A matrix of 0/1s, equal to Y != 0.}

\item{Y}{A data matrix of the same size as \code{V}.}

\item{parametrization}{A string, either \code{"abk"} (canonical) or \code{"pms"} (moment).}

\item{verbose}{A logical, whether to print intermediate steps.}

\item{control}{A list passed to \code{zi_fit()}. Please consult \code{?zi_fit}.}
}
\value{
An adjacency matrix.
}
\description{
Exhaustive search using BIC for DAGs for zero-inflated data based on Hurdle conditionals.
}
\details{
Performs exhaustive DAG search (using BIC) for DAGs for zero-inflated data based on Hurdle conditionals.
See Silander and Myllymaki (2006) or \code{?pcalg::simy} for details. However, unlike the Gaussian case in their implementation that returns an equivalence class of DAGs, the DAG estimated by this function is exact.
}
\examples{
m <- 4; n <- 1000
adj_mat <- ZiDAG::make_dag(m, mode = "chain", shuffle=TRUE)
d <- ZiDAG::gen_zero_dat(seed=1, gen_para="pms", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
est <- ZiDAG::ziSIMY(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=1L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est

d <- ZiDAG::gen_zero_dat(seed=1, gen_para="abk", adj_mat=adj_mat, n=n, gen_uniform_degree=1)
est <- ZiDAG::ziSIMY(V=d$V, Y=d$Y, parametrization="abk", verbose=FALSE,
    control=list(use_C=TRUE, maxit=1000, runs=2, report=0))
adj_mat == est

est <- ZiDAG::ziSIMY(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=1L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est

est <- ZiDAG::ziSIMY(V=d$V, Y=d$Y, parametrization="pms", verbose=FALSE,
    control=list("max_uniform_degree"=2L, "tol"=1e-8, "print_best_degree"=FALSE))
adj_mat == est
}
